requirements.txt
mysql-connector-python
sockets